Create Summary
Gales are a significant atmospheric issue that often occur during climatic changes, causing considerable damage to infrastructure, the environment, and living beings. Gale forecasting is a critical preventative measure to mitigate these impacts. In this context, we use XGBoost to predict whether a gale will occur.

XGBoost, or "Extreme Gradient Boosting," is an optimized distributed gradient-boosting library designed for efficient and scalable training of machine learning models. It is an ensemble learning method that combines the predictions of multiple weak models to produce a stronger and more accurate prediction.

This algorithm is renowned for its ability to handle large datasets and deliver state-of-the-art performance in various machine learning tasks, including classification and regression. Using XGBoost, we aim to create a model for future weather prediction, incorporating more than six critical parameters to determine whether a gale will occur.

The model leverages XGBoost's speed and efficiency to train data quickly and deliver predictions rapidly. Additionally, it is highly accurate, achieving a prediction accuracy of 92%, making it a valuable tool for weather forecasting and disaster prevention.